ported Avatar to the new structure
