+++ src/components/avatar/Avatar.props.ts
@@ -1,12 +1,27 @@
 import { TAvatarSizeToken, TAvatarStatus, TAvatarVariant } from './Avatar.types';
 
 type PAvatar = {
-    variant: TAvatarVariant;
     /**
      * the name to be displayed in the fallback avatar.
      * It gets capitalized and shortened to a maximum of 2 characters
      */
     name: string;
+    /**
+     * URL to fetch an avatar image.
+     */
+    image?: string;
+    /**
+     * Supported variants are:
+     *  'circle' - typically used in user avatars
+     *  'rounded' - typically used in team avatars
+     * @default 'circle'
+     */
+    variant?: TAvatarVariant;
+    /**
+     * the size token to define the Avatar size
+     * @default 'md'
+     */
+    size?: TAvatarSizeToken;
     /**
      * adds a hover effect to the avatar
      * @default false
@@ -25,15 +40,6 @@ type PAvatar = {
      */
     mentions?: number;
     /**
-     * URL to fetch an avatar image.
-     */
-    image?: string;
-    /**
-     * the size token to define the Avatar size
-     * @default 'md'
-     */
-    size?: TAvatarSizeToken;
-    /**
      * show a `StatusBadge` onm the bottom right of the avatar
      */
     status?: TAvatarStatus;
@@ -43,5 +49,10 @@ type PAvatar = {
     onClick: () => void;
     className?: string;
 };
+
+export type PAvatarRoot = Pick<
+    PAvatar,
+    'size' | 'variant' | 'disableHover' | 'isActive' | 'className'
+> & { hasUnreadBadge: boolean };
 
 export default PAvatar;
