+++ src/components/avatar/Avatar.stories.mdx
@@ -54,7 +54,7 @@ export const avatarArgtypes = {
 
 <Meta title={'Components/Avatar'} />
 
-<ArgsTable of={Avatar} exclude={[...DEFAULT_ARGUMENTSTABLE_EXCLUSION]} />
+<ArgsTable of={Avatar} />
 
 <Canvas hidden>
     <Story name="user avatar" args={userAvatarArgs} argTypes={avatarArgtypes}>
