+++ src/components/avatar/index.ts
@@ -1,7 +1,7 @@
-import Avatar from './Avatar';
+import AvatarRoot from './Avatar';
 
 export * from './Avatar.props';
 export * from './Avatar.types';
 export * from './Avatar.constants';
 
-export default Avatar;
+export default AvatarRoot;
